Added scripts for chocolatey packaging

+++ Tools/Setup/ConfigurationManagerSetup.ps1
@@ -3,11 +3,21 @@ param(
     [string] $installationRoot
 )
 
-Write-Host 'Stopping Service Jarvis - Configuration Service'
+if(!(Test-Path -Path $deployFileName ))
+{
+     Throw "Unable to find package file $deployFileName"
+}
 
+$file = Get-Item $deployFileName
+
+Write-Output "Starting Configuration Service installation script"
+Write-Output "deployFileName:  $deployFileName"
+Write-Output "installationRoot: $installationRoot"
+
+Write-Output 'Stopping Service Jarvis - Configuration Service'
 $configurationDir = "$installationRoot\ConfigurationStore"
 $installationRoot = "$installationRoot\ConfigurationManagerHost"
-Write-Host "Installing configuration manager host in $installationRoot"
+Write-Output "Installing configuration manager host in $installationRoot"
 $service = Get-Service -Name "Jarvis - Configuration Service" -ErrorAction SilentlyContinue 
 if ($service -ne $null) 
 {
@@ -15,24 +25,24 @@ if ($service -ne $null)
     $service.WaitForStatus("Stopped")
 }
 
-Write-Host 'Deleting actual directory'
+Write-Output 'Deleting actual directory'
 
 if (Test-Path $installationRoot) 
 {
     Remove-Item $installationRoot -Recurse -Force
 }
 
-Write-Host "Unzipping setup file"
-Expand-WithShell -zipFile $deployFileName -destinationFolder $installationRoot
+Write-Output "Unzipping setup file"
+Expand-WithShell -zipFile $file.FullName -destinationFolder $installationRoot
 
 if ($service -eq $null) 
 {
-    Write-Host "Starting the service in $finalInstallDir\Jarvis.ConfigurationService.Host.exe"
+    Write-Output "Starting the service in $finalInstallDir\Jarvis.ConfigurationService.Host.exe"
 
     & "$installationRoot\Jarvis.ConfigurationService.Host.exe" install
 } 
 
-Write-Host "Changing configuration"
+Write-Output "Changing configuration"
 
 $configFileName = $installationRoot + "\Jarvis.ConfigurationService.Host.exe.config"
 $xml = [xml](Get-Content $configFileName)
@@ -47,13 +57,20 @@ if(!(Test-Path -Path $configurationDir))
     New-Item -ItemType directory -Path $configurationDir
 }
 
-Write-Host "Removing sample directories"
+Write-Output "Removing sample directories"
 $configurationSampleDir = $installationRoot + "\Configuration.Sample"
-Remove-Item -Force -Path $configurationSampleDir -Recurse
-$configurationStoreDefaultDir = $installationRoot + "\ConfigurationStore"
-Remove-Item -Force -Path $configurationStoreDefaultDir -Recurse
+if (Test-Path ($configurationSampleDir)) 
+{
+    Remove-Item -Force -Path $configurationSampleDir -Recurse
+}
 
-Write-Host 'Starting the service'
+$configurationStoreDefaultDir = $installationRoot + "\ConfigurationStore"
+if (Test-Path ($configurationStoreDefaultDir)) 
+{
+    Remove-Item -Force -Path $configurationStoreDefaultDir -Recurse
+}
+
+Write-Output 'Starting the service'
 Start-Service "Jarvis - Configuration Service"
-Write-Host "Jarvis Configuration Service Installed"
+Write-Output "Jarvis Configuration Service Installed"
